requirements.txt
requests
python-freshdesk